requirements.txt
coverage==4.5.2
PyJWT==1.7.1
